feat: Add leaveGroup method to GroupController
This commit adds a new method, leaveGroup, to the GroupController. The leaveGroup method allows users to leave a group they are a member of. If the user is an admin (guide) of the group, they cannot leave their own group and an appropriate error message is returned. Otherwise, the user is detached from the group members and a success message is returned.